Fix bug where an alert for attacks would incorrectly not be sent
We ignore certain settlements when alerting for a period of time after user first alerted. For raids, there is an event on start and on end. Before, we only checked whether a settlement's ignore period had expired and removed it in the raiding-has-ended event (handled in FinalizeVillageRaid). This means that if a raid ended before the end of the ignoring period, the settlement would still appear in the list of ignored. Then, the next time a raid began, it would be ignored no matter what, regardless of whether the time period of ignoring had actually elapsed or not. This fixes that by additionally checking for expirations in the ignore list when the raiding-has-begun event fires (handled in DisplayVillageRaid).